+++ src/main/java/edu/metrostate/booknow/Controllers/LoginController.java
@@ -1,8 +1,7 @@
 package edu.metrostate.booknow.Controllers;
 
 import edu.metrostate.booknow.Services.UserServices;
-import edu.metrostate.booknow.Utils.AlertUtil;
-import edu.metrostate.booknow.Utils.SwitchSceneUtil;
+import edu.metrostate.booknow.Utils.UIUtils;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.PasswordField;
@@ -13,7 +12,6 @@ import java.io.IOException;
 /**
  * Controller class for handling login functionality in a JavaFX application.
  */
-
 public class LoginController {
 
     private static final String bookNowViewPath = "/edu/metrostate/booknow/BookNowView.fxml";
@@ -36,25 +34,14 @@ public class LoginController {
 
         // Move field validation to UserServices
         if (userService.areLoginFieldsValid(username, password)) {
-            if (userService.login(username, password)) {
-                switchToBookNowScene(event);
-            } else {
-                AlertUtil.showErrorAlert("Login Failed", "Invalid username or password.");
-            }
+            boolean loginSuccess = userService.login(username, password);
+            UIUtils.handleLogin(event, loginSuccess, bookNowViewPath, "Invalid username or password.");
         } else {
-            AlertUtil.showInfoAlert("Validation Error", "Both fields are required!");
+            UIUtils.showInfoAlert("Validation Error", "Both fields are required!");
         }
     }
 
     public void onCreateAccountButtonAction(ActionEvent event) throws IOException {
-        SwitchSceneUtil.switchScene(getClass().getResource(createAccountViewPath), event);
-    }
-
-    private void switchToBookNowScene(ActionEvent event) {
-        try {
-            SwitchSceneUtil.switchScene(getClass().getResource(bookNowViewPath), event);
-        } catch (IOException e) {
-            AlertUtil.showErrorAlert("Error", e.getMessage());
-        }
+        UIUtils.switchScene(getClass().getResource(createAccountViewPath), event);
     }
 }