+++ src/main/java/module-info.java
@@ -3,6 +3,7 @@ module edu.metrostate.booknow {
     requires javafx.fxml;
     requires java.sql;
     requires java.desktop;
+    requires jdk.jshell;
 
 
     opens edu.metrostate.booknow to javafx.fxml;